+++ tooling/scripts/src/commands/packages/format.ts
@@ -53,7 +53,6 @@ async function formatJson(json: string) {
 
 /** While we strive to make everything ESM, we just accept that some packages aren't ESM. */
 const NO_MODULE_PACKAGES = [
-  'scalar-app',
   '@scalar/docusaurus',
   '@scalar-examples/docusaurus',
   '@scalar-examples/nestjs-api-reference-express',