+++ integrations/dotnet/aspnetcore/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/aspnetcore
 
+## 2.13.1
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.48.0**
+  - [#8266](https://github.com/scalar/scalar/pull/8266): support operation level authentication and servers
+
 ## 2.13.0
 
 ### Minor Changes