+++ integrations/java/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/java-integration
 
+## 0.6.1
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.48.0**
+  - [#8266](https://github.com/scalar/scalar/pull/8266): support operation level authentication and servers
+
 ## 0.6.0
 
 ### Minor Changes