+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/api-client-react
 
+## 1.4.1
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-client@2.33.0**
+  - [#8266](https://github.com/scalar/scalar/pull/8266): support operation level authentication and servers
+
 ## 1.4.0
 
 ### Minor Changes